api/requirements.txt
cffi>=1.12.0,<=1.15.0
Django>=3.2.0,<=4.0.2
djangorestframework>=3.3.0,<=3.13.1
django-environ>=0.1.0,<=0.8.1
django-filter>=1.0.1,<=21.1
django-autoslug>1.6.0,<=1.9.8
django-countries>=5.0.0,<=7.2.1
django-phonenumber-field>=4.0.0,<=6.0.0
phonenumbers>=8.12.20,<=8.12.43
psycopg2>=2.6.0,<=2.9.3
gunicorn>=20.0.0,<=20.0.4
django-cors-headers>=3.3.0,<=3.10.0
uWSGI>=2.0.0,<=2.0.20
django-debug-toolbar>=3.0.0,<=3.2.4
djoser>=1.0.0,<=2.1.0
djangorestframework_simplejwt>=4.0.0,<=5.1.0
Fabric3>=1.14.post1
django-rest-swagger>=2.0.1,<=2.2.0
django-templated-email>=2.0.0,<=3.0.0
celery>=4.4.0,<=5.2.6
redis>=3.5.3,<=4.2.2
flower>=0.8.0,<=1.0.0
google-api-python-client>=2.26.0,<=2.49.0
google-auth>=2.3.0,<=2.6.6
google-auth-oauthlib>0.4.0,<=0.5.1
google-auth-httplib2>0.0.1,<=0.1.0
channels>3.0.0,<=3.0.5
channels-redis>3.0.0,<=3.4.1
stripe>3.0.0,<=4.1.0